# perf(ecs): Narrowing the cache search for the ECS provider on views