--- a/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/cache/client/EcsCloudWatchAlarmCacheClient.java
+++ b/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/cache/client/EcsCloudWatchAlarmCacheClient.java
@@ -73,6 +73,8 @@ public class EcsCloudWatchAlarmCacheClient extends AbstractCacheClient<EcsMetric
   public List<EcsMetricAlarm> getMetricAlarms(
       String serviceName, String accountName, String region) {
     List<EcsMetricAlarm> metricAlarms = new LinkedList<>();
+    // we can filter more here.
+
     Collection<EcsMetricAlarm> allMetricAlarms = getAll(accountName, region);
 
     outLoop: