--- a/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/deploy/ops/DestroyServiceAtomicOperation.java
+++ b/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/deploy/ops/DestroyServiceAtomicOperation.java
@@ -45,7 +45,10 @@ public class DestroyServiceAtomicOperation
 
     updateTaskStatus("Removing MetricAlarms from " + description.getServerGroupName() + ".");
     ecsCloudMetricService.deleteMetrics(
-        description.getServerGroupName(), description.getAccount(), description.getRegion());
+        description.getServerGroupName(),
+        description.getAccount(),
+        description.getRegion(),
+        ecsClusterName);
     updateTaskStatus("Done removing MetricAlarms from " + description.getServerGroupName() + ".");
 
     UpdateServiceRequest updateServiceRequest = new UpdateServiceRequest();