--- a/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/model/EcsApplication.java
+++ b/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/model/EcsApplication.java
@@ -26,11 +26,16 @@ public class EcsApplication implements Application {
   private String name;
   Map<String, String> attributes;
   Map<String, Set<String>> clusterNames;
+  Map<String, Set<String>> clusterNameMetadata;
 
   public EcsApplication(
-      String name, Map<String, String> attributes, Map<String, Set<String>> clusterNames) {
+      String name,
+      Map<String, String> attributes,
+      Map<String, Set<String>> clusterNames,
+      Map<String, Set<String>> getClusterNameMetadata) {
     this.name = name;
     this.attributes = attributes;
     this.clusterNames = clusterNames;
+    this.clusterNameMetadata = getClusterNameMetadata;
   }
 }