--- a/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/provider/agent/EcsCloudMetricAlarmCachingAgent.java
+++ b/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/provider/agent/EcsCloudMetricAlarmCachingAgent.java
@@ -118,7 +118,13 @@ public class EcsCloudMetricAlarmCachingAgent implements CachingAgent, AccountAwa
     Map<String, Collection<CacheData>> newDataMap = new HashMap<>();
 
     for (MetricAlarm metricAlarm : cacheableMetricAlarm) {
-      String key = Keys.getAlarmKey(accountName, region, metricAlarm.getAlarmArn());
+      String cluster =
+          metricAlarm.getDimensions().stream()
+              .filter(t -> t.getName().equalsIgnoreCase("ClusterName"))
+              .map(t -> t.getValue())
+              .collect(Collectors.joining());
+
+      String key = Keys.getAlarmKey(accountName, region, metricAlarm.getAlarmArn(), cluster);
       Map<String, Object> attributes =
           convertMetricAlarmToAttributes(metricAlarm, accountName, region);