--- a/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/provider/view/EcsServerClusterProvider.java
+++ b/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/provider/view/EcsServerClusterProvider.java
@@ -120,8 +120,13 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
       AmazonCredentials.AWSRegion awsRegion,
       String application) {
 
-    Collection<Service> services =
-        serviceCacheClient.getAll(credentials.getName(), awsRegion.getName());
+    String glob =
+        application != null
+            ? Keys.getServiceKey(credentials.getName(), awsRegion.getName(), application + "*")
+            : Keys.getServiceKey(credentials.getName(), awsRegion.getName(), "*");
+
+    Collection<String> ecsServices = serviceCacheClient.filterIdentifiers(glob);
+    Collection<Service> services = serviceCacheClient.getAll(ecsServices);
     Collection<Task> allTasks = taskCacheClient.getAll(credentials.getName(), awsRegion.getName());
 
     for (Service service : services) {
@@ -456,18 +461,12 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
 
   @Override
   public Map<String, Set<EcsServerCluster>> getClusterSummaries(String application) {
-    return getClusters();
+    return getClusters0(application);
   }
 
   @Override
   public Map<String, Set<EcsServerCluster>> getClusterDetails(String application) {
-    Map<String, Set<EcsServerCluster>> clusterMap = new HashMap<>();
-
-    for (AmazonCredentials credentials : getEcsCredentials()) {
-      clusterMap = findClusters(clusterMap, credentials, application);
-    }
-
-    return clusterMap;
+    return getClusters0(application);
   }
 
   @Override
@@ -476,6 +475,15 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
 
     for (AmazonCredentials credentials : getEcsCredentials()) {
       clusterMap = findClusters(clusterMap, credentials);
+    }
+    return clusterMap;
+  }
+
+  public Map<String, Set<EcsServerCluster>> getClusters0(String application) {
+    Map<String, Set<EcsServerCluster>> clusterMap = new HashMap<>();
+
+    for (AmazonCredentials credentials : getEcsCredentials()) {
+      clusterMap = findClusters(clusterMap, credentials, application);
     }
     return clusterMap;
   }