--- a/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/provider/view/EcsServerClusterProvider.java
+++ b/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/provider/view/EcsServerClusterProvider.java
@@ -100,15 +100,17 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
 
   private Map<String, Set<EcsServerCluster>> findClusters(
       Map<String, Set<EcsServerCluster>> clusterMap, AmazonCredentials credentials) {
-    return findClusters(clusterMap, credentials, null);
+    return findClusters(clusterMap, credentials, null, true);
   }
 
   private Map<String, Set<EcsServerCluster>> findClusters(
       Map<String, Set<EcsServerCluster>> clusterMap,
       AmazonCredentials credentials,
-      String application) {
+      String application,
+      boolean inludeDetails) {
     for (AmazonCredentials.AWSRegion awsRegion : credentials.getRegions()) {
-      clusterMap = findClustersForRegion(clusterMap, credentials, awsRegion, application);
+      clusterMap =
+          findClustersForRegion(clusterMap, credentials, awsRegion, application, inludeDetails);
     }
 
     return clusterMap;
@@ -118,7 +120,8 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
       Map<String, Set<EcsServerCluster>> clusterMap,
       AmazonCredentials credentials,
       AmazonCredentials.AWSRegion awsRegion,
-      String application) {
+      String application,
+      boolean includeDetails) {
 
     String glob =
         application != null
@@ -172,7 +175,8 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
               service.getClusterName(),
               taskDefinition,
               service.getSubnets(),
-              service.getSecurityGroups());
+              service.getSecurityGroups(),
+              includeDetails);
 
       if (ecsServerGroup == null) {
         continue;
@@ -313,13 +317,9 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
       String ecsClusterName,
       com.amazonaws.services.ecs.model.TaskDefinition taskDefinition,
       List<String> eniSubnets,
-      List<String> eniSecurityGroups) {
+      List<String> eniSecurityGroups,
+      boolean includeDetails) {
     ServerGroup.InstanceCounts instanceCounts = buildInstanceCount(instances);
-    TaskDefinition ecsTaskDefinition = buildTaskDefinition(taskDefinition);
-    EcsServerGroup.Image image = new EcsServerGroup.Image();
-    image.setImageId(ecsTaskDefinition.getContainerImage());
-    image.setName(ecsTaskDefinition.getContainerImage());
-
     String scalableTargetId = "service/" + ecsClusterName + "/" + serviceName;
     String scalableTargetKey = Keys.getScalableTargetKey(account, region, scalableTargetId);
     ScalableTarget scalableTarget = scalableTargetCacheClient.get(scalableTargetKey);
@@ -373,31 +373,52 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
         }
       }
     }
-
     Set<String> metricAlarmNames =
-        ecsCloudWatchAlarmCacheClient.getMetricAlarms(serviceName, account, region).stream()
+        ecsCloudWatchAlarmCacheClient
+            .getMetricAlarms(serviceName, account, region, ecsClusterName)
+            .stream()
             .map(EcsMetricAlarm::getAlarmName)
             .collect(Collectors.toSet());
-
-    EcsServerGroup serverGroup =
-        new EcsServerGroup()
-            .setDisabled(capacity.getDesired() == 0)
-            .setName(serviceName)
-            .setCloudProvider(EcsCloudProvider.ID)
-            .setType(EcsCloudProvider.ID)
-            .setRegion(region)
-            .setInstances(instances)
-            .setCapacity(capacity)
-            .setImage(image)
-            .setInstanceCounts(instanceCounts)
-            .setCreatedTime(creationTime)
-            .setEcsCluster(ecsClusterName)
-            .setTaskDefinition(ecsTaskDefinition)
-            .setVpcId(vpcId)
-            .setSecurityGroups(securityGroups)
-            .setMetricAlarms(metricAlarmNames)
-            .setMoniker(moniker);
-
+    EcsServerGroup serverGroup = new EcsServerGroup();
+    if (includeDetails) {
+      TaskDefinition ecsTaskDefinition = buildTaskDefinition(taskDefinition);
+      EcsServerGroup.Image image = new EcsServerGroup.Image();
+      image.setImageId(ecsTaskDefinition.getContainerImage());
+      image.setName(ecsTaskDefinition.getContainerImage());
+      serverGroup
+          .setDisabled(capacity.getDesired() == 0)
+          .setName(serviceName)
+          .setCloudProvider(EcsCloudProvider.ID)
+          .setType(EcsCloudProvider.ID)
+          .setRegion(region)
+          .setInstances(instances)
+          .setCapacity(capacity)
+          .setImage(image)
+          .setInstanceCounts(instanceCounts)
+          .setCreatedTime(creationTime)
+          .setEcsCluster(ecsClusterName)
+          .setTaskDefinition(ecsTaskDefinition)
+          .setVpcId(vpcId)
+          .setSecurityGroups(securityGroups)
+          .setMetricAlarms(metricAlarmNames)
+          .setMoniker(moniker);
+    } else {
+      serverGroup
+          .setDisabled(capacity.getDesired() == 0)
+          .setName(serviceName)
+          .setCloudProvider(EcsCloudProvider.ID)
+          .setType(EcsCloudProvider.ID)
+          .setRegion(region)
+          .setInstances(instances)
+          .setCapacity(capacity)
+          .setInstanceCounts(instanceCounts)
+          .setCreatedTime(creationTime)
+          .setEcsCluster(ecsClusterName)
+          .setVpcId(vpcId)
+          .setSecurityGroups(securityGroups)
+          .setMetricAlarms(metricAlarmNames)
+          .setMoniker(moniker);
+    }
     EcsServerGroup.AutoScalingGroup asg =
         new EcsServerGroup.AutoScalingGroup()
             .setDesiredCapacity(scalableTarget.getMaxCapacity())
@@ -461,12 +482,12 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
 
   @Override
   public Map<String, Set<EcsServerCluster>> getClusterSummaries(String application) {
-    return getClusters0(application);
+    return getClusters0(application, false);
   }
 
   @Override
   public Map<String, Set<EcsServerCluster>> getClusterDetails(String application) {
-    return getClusters0(application);
+    return getClusters0(application, true);
   }
 
   @Override
@@ -479,11 +500,12 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
     return clusterMap;
   }
 
-  public Map<String, Set<EcsServerCluster>> getClusters0(String application) {
+  private Map<String, Set<EcsServerCluster>> getClusters0(
+      String application, boolean includeDetails) {
     Map<String, Set<EcsServerCluster>> clusterMap = new HashMap<>();
 
     for (AmazonCredentials credentials : getEcsCredentials()) {
-      clusterMap = findClusters(clusterMap, credentials, application);
+      clusterMap = findClusters(clusterMap, credentials, application, includeDetails);
     }
     return clusterMap;
   }
@@ -493,7 +515,7 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
   public Set<EcsServerCluster> getClusters(String application, String account) {
     try {
       AmazonCredentials credentials = getEcsCredentials(account);
-      return findClusters(new HashMap<>(), credentials, application).get(application);
+      return findClusters(new HashMap<>(), credentials, application, true).get(application);
     } catch (NoSuchElementException exception) {
       log.info("No ECS Credentials were found for account " + account);
       return null;
@@ -544,7 +566,7 @@ public class EcsServerClusterProvider implements ClusterProvider<EcsServerCluste
       AmazonCredentials credentials = getEcsCredentials(account);
       Moniker moniker = MonikerHelper.applicationNameToMoniker(serverGroupName);
       log.debug("App Name is: " + moniker.getApp());
-      clusterMap = findClusters(clusterMap, credentials, moniker.getApp());
+      clusterMap = findClusters(clusterMap, credentials, moniker.getApp(), true);
     } catch (NoSuchElementException exception) {
       /* This is ugly, but not sure how else to do it. If we don't have creds due
        *  to not being an ECS account, there's nothing to do here, and we should