--- a/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/services/EcsCloudMetricService.java
+++ b/clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/services/EcsCloudMetricService.java
@@ -43,9 +43,10 @@ public class EcsCloudMetricService {
 
   private final Logger log = LoggerFactory.getLogger(getClass());
 
-  public void deleteMetrics(String serviceName, String account, String region) {
+  public void deleteMetrics(
+      String serviceName, String account, String region, String ecsClusterName) {
     List<EcsMetricAlarm> metricAlarms =
-        metricAlarmCacheClient.getMetricAlarms(serviceName, account, region);
+        metricAlarmCacheClient.getMetricAlarms(ecsClusterName, serviceName, account, region);
 
     if (metricAlarms.isEmpty()) {
       return;