--- a/clouddriver-web/src/main/groovy/com/netflix/spinnaker/clouddriver/controllers/ClusterController.groovy
+++ b/clouddriver-web/src/main/groovy/com/netflix/spinnaker/clouddriver/controllers/ClusterController.groovy
@@ -16,6 +16,7 @@
 
 package com.netflix.spinnaker.clouddriver.controllers
 
+import com.netflix.spinnaker.clouddriver.ecs.model.EcsApplication
 import com.netflix.spinnaker.clouddriver.model.*
 import com.netflix.spinnaker.clouddriver.model.view.ClusterViewModelPostProcessor
 import com.netflix.spinnaker.clouddriver.model.view.ServerGroupViewModelPostProcessor
@@ -85,9 +86,12 @@ class ClusterController {
 
   private Map<String, Set<String>> mergeClusters(List<Application> a) {
     Map<String, Set<String>> map = new HashMap<>()
-
     a.stream()
-      .flatMap({ it.getClusterNames().entrySet().stream() })
+      .flatMap({
+        it instanceof EcsApplication
+        ? it.getClusterNameMetadata().entrySet().stream()
+        : it.getClusterNames().entrySet().stream()
+      })
       .forEach({ entry ->
         map.computeIfAbsent(entry.getKey(), { new HashSet<>() }).addAll(entry.getValue())
       })