Attempt to address some of the issues described in https://github.com/spinnaker/spinnaker/issues/6084

Improving the response times on:
- /clusters
- /applications
- /serverGroups
Endpoints when ECS is enabled and a substantial amount of accounts/services exist in cache.

The perf issue with the Alarms still exists and will be addressed in a future PR


Adding some results from a performance test clouddriver response times:
![image](https://github.com/user-attachments/assets/f0f7a4d2-f00a-4ed0-8d59-a937f014bf0c)
![image](https://github.com/user-attachments/assets/d2bd74b1-45e9-43d1-8ebf-fd1b0afe7a4b)
![image](https://github.com/user-attachments/assets/e1c6cf4a-ce58-4124-bb5c-4061c61d0645)

- GET {CLOUDDRIVER_URL}/applications
    - Average: 104ms → 92.1ms (11% improvement)
    - 95th Percentile: 130ms → 126ms
- GET {CLOUDDRIVER_URL}/applications/{application_name}
  - Average: 7.48s → 4.2s (43% improvement)
  - 95th Percentile: 8.55s → 5.86s
- GET {CLOUDDRIVER_URL}/applications/{application_name}/serverGroups
  - Average: 2.72s → 2.16s (20% improvement)
  - 95th Percentile: 3.17s → 3.11s
- GET {CLOUDDRIVER_URL}/applications/{application_name}/clusters
  - Average: 107ms → 43.3ms (59% improvement)
  - 95th Percentile: 135ms → 88.4ms