@christosarvanitis 2024-09-04
@dbyron-sf @jasonmcintosh Added some results from an internal testing related to this change. Would appreciate any feedback!