@christosarvanitis 2024-09-20
Following up with the Alarm changes the perf improvements:
![image](https://github.com/user-attachments/assets/37ebb4d9-25b0-45c4-a479-379b8d821ddd)

The GET ServerGroups call reduced from 7+ secs on an application to 3.5secs.

And timings on a single request before the change:
![image](https://github.com/user-attachments/assets/baf2e6c1-018a-414f-b9ad-fd095576484a)
And after the change:
![image](https://github.com/user-attachments/assets/764dc701-4428-4cfd-ab91-cb0a52843538)