@jasonmcintosh 2024-09-20
Overall I think this looks good ;) Would like one more set of eyes given the changes on the way ECS operates.  One concern is how the change on the cache id's will be cleaned up since this changes the storage ids - but that MAY get taken care of by one of the cleanup jobs (need to confirm).  Talked in slack - the ID's should still fit the max column length so adding that shouldn't impact any database stuff.